improvde docs + remove inventory feature from equipment
